+++ src/atlas.rs
@@ -12,27 +12,7 @@ pub use paths::get_exe_folder;
 mod point;
 pub use point::{Point, SizedPoint};
 
+mod easy_path;
+pub use easy_path::EasyPath;
+
 pub type BoxResult<T> = Result<T, Box<dyn std::error::Error>>;
-
-pub trait EasyPath {
-    fn stringify(&self) -> &str;
-    fn stringify_owned(&self) -> String;
-}
-
-impl EasyPath for std::path::Path {
-    fn stringify(&self) -> &str {
-        self.to_str().unwrap()
-    }
-    fn stringify_owned(&self) -> String {
-        self.stringify().to_string()
-    }
-}
-
-impl EasyPath for std::path::PathBuf {
-    fn stringify(&self) -> &str {
-        self.to_str().unwrap()
-    }
-    fn stringify_owned(&self) -> String {
-        self.stringify().to_string()
-    }
-}
